backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
playwright==1.48.0
python-multipart==0.0.17
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.54.0
supabase==2.10.0
python-dotenv==1.0.1
